Struct substrate_client_db::Backend

pub struct Backend<Block: BlockT> { /* fields omitted */ }

Disk backend. Keeps data in a key-value store. In archive mode, trie nodes are kept from all blocks. Otherwise, trie nodes are kept only from some recent blocks.

Methods

impl<Block: BlockT<Hash = H256>> Backend<Block>

pub fn new(
config: DatabaseSettings,
canonicalization_delay: u64
) -> Result<Self, Error>

Create a new instance of database backend.

The pruning window is how old a block must be before the state is pruned.

pub fn new_test(keep_blocks: u32, canonicalization_delay: u64) -> Self

pub fn as_in_memory(&self) -> InMemoryBackend<Block, Blake2Hasher>

Returns in-memory blockchain that contains the same set of blocks that the self.